feat: add new backend `/exists` for #114
We are aware that this is a a potential security risks, because it allows
to check if a user is registered and then
target password attacks.
However, we'd like to simplify the login process for students and
and forgo passwords for them.
Hence we'll have to add rate-limiting on this API